Добавил расчет изменения баланса, изменил запросы в репозиториях, чтобы возвращали 1 ответ и изменил тип для баланса на лонг

+++ src/main/java/com/application/billingsystem/dto/PaymentCreateDto.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class PaymentCreateDto {
     private String numberPhone;
-    private float money;
+    private long money;
 }
